requirements.txt
ultralytics
matplotlib
Pillow
streamlit
